vcl: protect more outdev functions for disposed state
This is similar to commit c612c3b0aed9ad7f7f42b4313f821b71995ead15
(protect more printer code-paths., 2015-03-20), but handles more
OutputDevice member functions.

The user-level problem was that in case a macro creates a dialog with an
embedded Chart document and the user clicks on e.g. the chart title (so
an sdr::overlay::OverlayManager is created), then it can happen during
closing the dialog that the overlay manager calls these functions after
the output device is disposed.

Change-Id: I8021fb795704f19e52d70505804d68725c636ce0
Reviewed-on: https://gerrit.libreoffice.org/66403
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins