tdf#122238 KDE5: disable mechanism of Qt for shortcut activation
When QEvent::KeyPress event is processed, QEvent::ShortcutOverride event is created
and processed before finishing processing original event.

If it's an active shortcut, and QEvent::ShortcutOverride event is not accepted,
the shortcut is activated, but QEvent::KeyPress event is omitted.
If it's an active shortcut, and QEvent::ShortcutOverride event is accepted,
the shortcut is not activated, and QEvent::KeyPress event is sent as usual.

But it looks like even if shortcut activation is suppressed,
shortcut is still called somewhere from insides SalEvent::KeyInput
event processing callback from Qt5Widget::handleKeyEvent function.

Shortcuts are still assigned to menu actions
since it shows the key combinations used to activate them.

Change-Id: I1ffc63c2530dd61ae67ab66a96ea3efe767218c0
Reviewed-on: https://gerrit.libreoffice.org/66444
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>