tdf#122878: enable wrap for flys in footer
This patch removes the check for a footer node,
intoduced by
23f698ecee033612ac3a9f5cfd7674b08bb3ccd1
preserving the behaviour for the connected bug reports
i13832 and i24135.

Without this check, the wraping becomes enabled for footer
objects, too.

With this enhencement, the commits
7df33caac85ac90c26e97dedbc201f46dc9e4cb4
d3db6ff43a531ecf1afc858a0a8832353d091644
are directly affected and therefore the unit test is edited.

Change-Id: I093add9e251ac97859a66cb8b8563010ef734c2d
Reviewed-on: https://gerrit.libreoffice.org/67069
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>