fix load/save of forms with HAVING
Since
    commit 2b1d6f0d3b0b025148c81986ba7f109659d838af
    Date:   Sun Jul 30 17:57:14 2017 +0200
    tdf#96370 rework filtering to be aware of WHERE vs HAVING clause

(1) the code was saving PROPERTY_FILTER and PROPERTY_SORT, but loading
PROPERTY_FILTER and PROPERTY_HAVINGCLAUSE
(2) the loading code was not putting the properties into m_xAggregateSet

So I fixed the load/save mismatch, put the properties into the right
location, and added the HAVING property, which meant I needed to do a
version bump.

I will note that this is somewhat of a backwards/forwards compatibility
mess - the commit referenced above added new fields in the middle of the
output stream, which means that older versions of LO will not be able
to properly load forms generated after that commit.

Change-Id: I9a13877b29d7c6bc5e6d014cfbcefd3069ddc4b5
Reviewed-on: https://gerrit.libreoffice.org/66830
Tested-by: Jenkins
Reviewed-by: Lionel Elie Mamane <lionel@mamane.lu>