sw: make it possible to listen to paste events
The use-case is that the user pastes content and custom postprocessing
is necessary for the pasted content. This is not easy by default, since
the cursor is at the end of the pasted content, so the start is not
known -- even if you intercept the paste UNO command.

Precisely tracking the start is possible when using the internal API,
SwReader::Read() does something similar already to track the undo range.

Only expose the text selection after paste (as a start), but design the
API in a way so that other pasted content type (e.g. image) can be
added in a backwards-compatible way later.

Change-Id: If35bda4244cc417e4858124fb75f7608b95fc556
Reviewed-on: https://gerrit.libreoffice.org/66807
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins