CppunitTest_desktop_lib: avoid use-after-free
The LOK callback registers a DesktopLOKTest instance, but depending on
timing, VCL may invoke that callback after that instance is already
deleted. The symptom is that all tests run fine, but the testsuite still
crashes at the end.

This can happen because DeInitVCL() indirectly calls
Scheduler::ProcessEventsToIdle(), which
desktop::CallbackFlushHandler::Invoke(), which may or may not invoke the
(now gone) callback, depending on timing. This typically doesn't happen
in debug builds, but it happens in release builds.

Change-Id: I00fc07046c46ad602a8b26c022ffcc916d211346
Reviewed-on: https://gerrit.libreoffice.org/66909
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins