tdf#98620 filter\ww8 export: spam Jc if environment defines BiDi
If the BiDi value value comes from the page style, then MS formats
have no idea what to do with it, so those values are written
out into the paragraph itself. Since Justify is highly dependent
on BiDi in order to understand its meaning, it also needs to
be spammed.

Change-Id: I7407056573bb115e8bab2dce0070b0a718dcc1eb
Reviewed-on: https://gerrit.libreoffice.org/66923
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>