add PngImageFilter that uses libpng for PNG loading
This adds loading of PNG images that uses libpng instead of our
own solution. It always loaded the image as either RGB or RGBA
image and if the source PNG is using something else, libpng
converts to either RGB or RGBA.

In addition this adds tests for loading of various PNG files to
make sure the resulting bitmaps are using pixel data as expected.
(especially needed to check the RGBA bitmaps)

Change-Id: I194321caf76c2ec2365bb6075c79c5e84983658a
Reviewed-on: https://gerrit.libreoffice.org/67571
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>