sw btlr writing mode: implement UNO API
The custom code in SwXCell::setPropertyValue() was added in commit
5a5597655a4bf12e4ca07c9c2b6f6221e217f080 (tentative fix for fdo#30474#,
2010-11-26), which suggests that not handling all constants from
text::WritingMode2 was not intentional.

Later the writerfilter side (which is the only client of this hidden
property) was adapted to use text::WritingMode2, so do the same here.
This implicitly adds support for the new text::WritingMode2::BT_LR as
well.

Change-Id: I37d8eaa844847cb19e7503b2d973069f9895e6bc
Reviewed-on: https://gerrit.libreoffice.org/67730
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins