tdf#123440 new expert option 'WriteLayerStateAsConfigItem'
LibreOffice writes the ODF attributes 'display' and 'protected' of
the 'layer' element since version 6.2. It can read the ODF
attributes since version 6.2. But they are only evaluated if non of
the config items VisibleLayers, PrintableLayers and LockedLayers
exists. If the option is set to 'false' no config items are written
and so reading of ODF attributes is forced. Version 6.1 is the last
one which needs the config items. So until version 6.1 is end of
life the default value is 'true'.
The patch introduces this option and adapts the relevant unit
tests to consider the option.

Change-Id: I2f8fc61506b14820f6bee8bf153d89363690b9e8
Reviewed-on: https://gerrit.libreoffice.org/67784
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>