tdf#123474: use INetURLObject methods to construct valid URI
This ensures that no extra slashes are added to the path when it
already ends with a slash; for Windows drive root path (e.g. C:\)
the trailing slash is always kept to not change the meaning to
"current path on the drive", which is different from "root of the
drive". Our IsValidFilePath does not allow more than one slash
after <drive:>.

Change-Id: Ife3cd9e146573a0c278834f795f0d7318c2d303a
Reviewed-on: https://gerrit.libreoffice.org/67850
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>
Tested-by: Stephan Bergmann <sbergman@redhat.com>