tdf#123451 qt5: Detect decimal separator on keypad
'QtKeyEvent::key()' doesn't return a special value for the
decimal separator key on the keypad ("," or "."), but just returns
'Qt::Key_Comma' or 'Qt::Key_Period'. However, the 'Qt::KeypadModifier'
modifier is set in this case, so check for this one in addition
and return 'KEY_DECIMAL' if those are combined.

Change-Id: Ia80826e2ad5e47a1f49bef450168523d766c1d6a
Reviewed-on: https://gerrit.libreoffice.org/67886
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>