tdf#116382 Replace old dialog in header and footer drop-down menu
Remove the Background tabpage in "Border/Background" dialog accessed
from Border and Background menu item in Header/Footer drop-down list
and replace it with the tabpages "Area" and "Transparency" of the dialog
accessed from Page Style > Header > "More.." button.
 The old dialog does not have the tabpages "Area" and "Transparency". 
In simpler words, the patch is to match both the dialogs
when accessed in two different ways.

Change-Id: Ib84435854389360eba41aecb8578a9a67a92ca1e
Reviewed-on: https://gerrit.libreoffice.org/67483
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>