We don't want any MODELESSDIALOG dialogs in the iOS app, they don't work
(And probably the same holds for Android, too.)

This change makes at least the Format>Character... and
Format>Paragraph... dialogs in the iOS app show up. For some reason
they are claimed to be modeless, and that meant that the mechanism for
dialog display in the iOS app did not work. The exact mecahnism is
unclear to me, but the alleged modelessness was the root cause.

Change-Id: I237a5e16ec34221912a2fb062068101562ba7dbb