Use comphelper::getProcessComponentContext()
The context in the mxContext member does not seem to be usable for
what we need it for here. Using what
comphelper::getProcessComponentContext() returns works better. Let's
hope it has no unintended side-effects.

This likely makes the mxContext member unused, but I did not bother
removing it yet, to keep this commit minimal.

Change-Id: Ic048683b066af7952e2e84b03ea306e7daaba259
Reviewed-on: https://gerrit.libreoffice.org/68592
Tested-by: Jenkins
Reviewed-by: Tor Lillqvist <tml@collabora.com>