tdf#124212 make adjustment handle work always
In method SetHandleControllerPosition() all transformations are
reversed to get the coordinates in shape inner coordinate system
from the outer position. The error was, that the first
transformation, the move in method GetPoint was forgotten. In case
of default viewBox '0 0 21600 21600' it is not visible. But the
error is noticeable, if left or top do not equal zero.

Change-Id: Icc3f4f2c603826151c95b8b9eea5030fb5805d67
Reviewed-on: https://gerrit.libreoffice.org/69439
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>