tdf#124109: Revert "Only do kashida insertion with fonts that have non-zero width kashidas"
This reverts commit c45b23377bb2fe44c26f1287ff38495344e4ca50.

This commit breaks Kashida justification when a document is first opened (the
kashida justification will be disabled and spaces will be used untill the text
is changed). Probably it is checking for width of Kashida glyph too early. Also
I'm not sure what the reverted commit was trying to fix since we already do
kashida justification in fonts that has non-zero width Kashida and it has been
the case for a long time.

This does not fix the original issue in the document attched with the bug
report, but if fixes kashida not being applied when opening the document.

Change-Id: Ic95859bca94fa792793e3223d2adb465bc6d880f
Reviewed-on: https://gerrit.libreoffice.org/69509
Tested-by: Jenkins
Reviewed-by: Xisco Faulí <xiscofauli@libreoffice.org>